packages/backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
python-dotenv==1.2.1
PyPDF2==3.0.1
openai==2.15.0
numpy==2.4.1
boto3==1.38.0
httpx==0.28.0
authlib==1.5.0
PyJWT==2.10.1